services/web_ui/pyproject.toml
[project]
name = "graph-rag-web-ui"
version = "0.1.0"
description = "Flask web UI for Graph RAG system"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.0.0",
    "neo4j>=5.0.0",
    "qdrant-client>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services"]

